Applied Gentoo patches from Mike Frysinger <vapier@gentoo.org> :
One of the regcomp calls was checked to make sure it worked, this patch adds
checking to the other call.

One of the malloc calls was checked to make sure it worked, this patch adds
checking to the other call ... in Gentoo we created a local xmalloc() which
called malloc and checked the result and used that instead so we didnt have
to add checks to every malloc call.

Seems the uclibc/nls hack was merged into eject cvs ... this was a temp hack
added to Gentoo but never meant for upstream :)  i've rewritten it into a
proper fix and attached here ... now you just run `./configure --disable-nls`

The -O3 hardcoded in Makefile.am should probably be removed ... if CFLAGS isnt
set in the environment when configure is run, then autoconf will select a
'proper' set of default CFLAGS.